move hook from utils to hooks dir
